# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  selectedCharacterId: string;
  difficultyTags: string[];
  xp: number;
  createdAt: string;
}

export interface CreateUserBody {
  name: string;
  selectedCharacterId?: string;
  difficultyTags?: string[];
}

export interface UpdateUserBody {
  name?: string;
  selectedCharacterId?: string;
  difficultyTags?: string[];
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  open: "open",
  scheduled: "scheduled",
  in_progress: "in_progress",
  paused: "paused",
  completed: "completed",
} as const;

export type TaskResistanceLevel =
  (typeof TaskResistanceLevel)[keyof typeof TaskResistanceLevel];

export const TaskResistanceLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: string;
  userId: string;
  title: string;
  status: TaskStatus;
  startTime?: string | null;
  endTime?: string | null;
  scheduledFor?: string | null;
  durationMinutes?: number | null;
  resistanceLevel: TaskResistanceLevel;
  taskType?: string | null;
  createdAt: string;
}

export type CreateTaskBodyResistanceLevel =
  (typeof CreateTaskBodyResistanceLevel)[keyof typeof CreateTaskBodyResistanceLevel];

export const CreateTaskBodyResistanceLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  userId: string;
  title: string;
  durationMinutes?: number | null;
  resistanceLevel?: CreateTaskBodyResistanceLevel;
  taskType?: string | null;
  scheduledFor?: string | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  open: "open",
  scheduled: "scheduled",
  in_progress: "in_progress",
  paused: "paused",
  completed: "completed",
} as const;

export type UpdateTaskBodyResistanceLevel =
  (typeof UpdateTaskBodyResistanceLevel)[keyof typeof UpdateTaskBodyResistanceLevel];

export const UpdateTaskBodyResistanceLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  status?: UpdateTaskBodyStatus;
  scheduledFor?: string | null;
  startTime?: string | null;
  endTime?: string | null;
  durationMinutes?: number | null;
  resistanceLevel?: UpdateTaskBodyResistanceLevel;
}

export interface ExtractTasksBody {
  rawText: string;
}

export type ExtractedTaskResistanceLevel =
  (typeof ExtractedTaskResistanceLevel)[keyof typeof ExtractedTaskResistanceLevel];

export const ExtractedTaskResistanceLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ExtractedTask {
  title: string;
  durationMinutes?: number | null;
  resistanceLevel: ExtractedTaskResistanceLevel;
  taskType?: string | null;
}

export interface ConfirmExtractedBody {
  userId: string;
  tasks: ExtractedTask[];
}

export interface ChatRequest {
  userId: string;
  characterId: string;
  message: string;
  sessionId?: string | null;
  activeTaskId?: string | null;
}

export interface UIBlock {
  type: string;
  data?: unknown;
  options?: string[];
  items?: string[];
  tasksCompleted?: string[];
  tasksPartial?: string[];
  xpEarned?: number;
}

export interface ChatResponse {
  message: string;
  uiBlocks: UIBlock[];
}

export interface ChatMessage {
  id: string;
  userId: string;
  characterId: string;
  sessionId?: string | null;
  role: string;
  content: string;
  uiBlocks: UIBlock[];
  createdAt: string;
}

export interface Session {
  id: string;
  userId: string;
  characterId: string;
  activeTaskId?: string | null;
  startTime: string;
  endTime?: string | null;
  summary?: string | null;
  createdAt: string;
}

export interface CreateSessionBody {
  userId: string;
  characterId: string;
  activeTaskId?: string | null;
}

export interface UpdateSessionBody {
  endTime?: string | null;
  summary?: string | null;
  activeTaskId?: string | null;
}

export interface AwardXpBody {
  userId: string;
  eventType: string;
  xpAwarded: number;
}

export type CharacterMode = (typeof CharacterMode)[keyof typeof CharacterMode];

export const CharacterMode = {
  chaos: "chaos",
  focus: "focus",
  calm: "calm",
  sarcasm: "sarcasm",
  paradox: "paradox",
  urgent: "urgent",
} as const;

export interface Character {
  id: string;
  name: string;
  mode: CharacterMode;
  xpRequired: number;
  color: string;
  personalityPrompt: string;
  tagline?: string;
  isLocked: boolean;
}

export interface AwardXpResponse {
  newXp: number;
  unlockedCharacter?: Character | null;
}

export type ListTasksParams = {
  userId: string;
  status?: string;
};

export type ExtractTasks200 = {
  tasks: ExtractedTask[];
};

export type GetChatHistoryParams = {
  userId: string;
  characterId?: string;
  limit?: number;
};

export type ListCharactersParams = {
  userId?: string;
};