# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spark — ADHD Companion API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: users
  - name: tasks
  - name: chat
  - name: sessions
  - name: xp
  - name: characters
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - { name: userId, in: query, required: true, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }

  /tasks/extract:
    post:
      operationId: extractTasks
      tags: [tasks]
      summary: Use AI to extract structured tasks from raw text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractTasksBody"
      responses:
        "200":
          description: Extracted task previews
          content:
            application/json:
              schema:
                type: object
                required: [tasks]
                properties:
                  tasks:
                    type: array
                    items: { $ref: "#/components/schemas/ExtractedTask" }

  /tasks/confirm-extract:
    post:
      operationId: confirmExtractedTasks
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmExtractedBody"
      responses:
        "201":
          description: Saved tasks
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message and receive AI response with optional UI blocks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Assistant message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      parameters:
        - { name: userId, in: query, required: true, schema: { type: string } }
        - { name: characterId, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: Recent chat messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }

  /sessions:
    post:
      operationId: createSession
      tags: [sessions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  /sessions/{sessionId}:
    patch:
      operationId: updateSession
      tags: [sessions]
      parameters:
        - { name: sessionId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  /xp/award:
    post:
      operationId: awardXp
      tags: [xp]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardXpBody"
      responses:
        "200":
          description: New XP and any unlocked characters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwardXpResponse"

  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      parameters:
        - { name: userId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: List of characters with unlock status
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Character" }

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    User:
      type: object
      required: [id, name, selectedCharacterId, difficultyTags, xp, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        selectedCharacterId: { type: string }
        difficultyTags:
          type: array
          items: { type: string }
        xp: { type: integer }
        createdAt: { type: string, format: date-time }

    CreateUserBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        selectedCharacterId: { type: string }
        difficultyTags:
          type: array
          items: { type: string }

    UpdateUserBody:
      type: object
      properties:
        name: { type: string }
        selectedCharacterId: { type: string }
        difficultyTags:
          type: array
          items: { type: string }

    Task:
      type: object
      required: [id, userId, title, status, resistanceLevel, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        title: { type: string }
        status:
          type: string
          enum: [open, scheduled, in_progress, paused, completed]
        startTime: { type: string, format: date-time, nullable: true }
        endTime: { type: string, format: date-time, nullable: true }
        scheduledFor: { type: string, format: date-time, nullable: true }
        durationMinutes: { type: integer, nullable: true }
        resistanceLevel:
          type: string
          enum: [low, medium, high]
        taskType: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }

    CreateTaskBody:
      type: object
      required: [userId, title]
      properties:
        userId: { type: string }
        title: { type: string }
        durationMinutes: { type: integer, nullable: true }
        resistanceLevel:
          type: string
          enum: [low, medium, high]
        taskType: { type: string, nullable: true }
        scheduledFor: { type: string, format: date-time, nullable: true }

    UpdateTaskBody:
      type: object
      properties:
        title: { type: string }
        status:
          type: string
          enum: [open, scheduled, in_progress, paused, completed]
        scheduledFor: { type: string, format: date-time, nullable: true }
        startTime: { type: string, format: date-time, nullable: true }
        endTime: { type: string, format: date-time, nullable: true }
        durationMinutes: { type: integer, nullable: true }
        resistanceLevel:
          type: string
          enum: [low, medium, high]

    ExtractTasksBody:
      type: object
      required: [rawText]
      properties:
        rawText: { type: string }

    ExtractedTask:
      type: object
      required: [title, resistanceLevel]
      properties:
        title: { type: string }
        durationMinutes: { type: integer, nullable: true }
        resistanceLevel:
          type: string
          enum: [low, medium, high]
        taskType: { type: string, nullable: true }

    ConfirmExtractedBody:
      type: object
      required: [userId, tasks]
      properties:
        userId: { type: string }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/ExtractedTask" }

    ChatRequest:
      type: object
      required: [userId, characterId, message]
      properties:
        userId: { type: string }
        characterId: { type: string }
        message: { type: string }
        sessionId: { type: string, nullable: true }
        activeTaskId: { type: string, nullable: true }

    UIBlock:
      type: object
      required: [type]
      properties:
        type: { type: string }
        data: {}
        options:
          type: array
          items: { type: string }
        items:
          type: array
          items: { type: string }
        tasksCompleted:
          type: array
          items: { type: string }
        tasksPartial:
          type: array
          items: { type: string }
        xpEarned: { type: integer }

    ChatResponse:
      type: object
      required: [message, uiBlocks]
      properties:
        message: { type: string }
        uiBlocks:
          type: array
          items: { $ref: "#/components/schemas/UIBlock" }

    ChatMessage:
      type: object
      required: [id, userId, characterId, role, content, uiBlocks, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        characterId: { type: string }
        sessionId: { type: string, nullable: true }
        role: { type: string }
        content: { type: string }
        uiBlocks:
          type: array
          items: { $ref: "#/components/schemas/UIBlock" }
        createdAt: { type: string, format: date-time }

    Session:
      type: object
      required: [id, userId, characterId, startTime, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        characterId: { type: string }
        activeTaskId: { type: string, nullable: true }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time, nullable: true }
        summary: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }

    CreateSessionBody:
      type: object
      required: [userId, characterId]
      properties:
        userId: { type: string }
        characterId: { type: string }
        activeTaskId: { type: string, nullable: true }

    UpdateSessionBody:
      type: object
      properties:
        endTime: { type: string, format: date-time, nullable: true }
        summary: { type: string, nullable: true }
        activeTaskId: { type: string, nullable: true }

    AwardXpBody:
      type: object
      required: [userId, eventType, xpAwarded]
      properties:
        userId: { type: string }
        eventType: { type: string }
        xpAwarded: { type: integer }

    AwardXpResponse:
      type: object
      required: [newXp]
      properties:
        newXp: { type: integer }
        unlockedCharacter:
          $ref: "#/components/schemas/Character"
          nullable: true

    Character:
      type: object
      required: [id, name, mode, xpRequired, color, personalityPrompt, isLocked]
      properties:
        id: { type: string }
        name: { type: string }
        mode:
          type: string
          enum: [chaos, focus, calm, sarcasm, paradox, urgent]
        xpRequired: { type: integer }
        color: { type: string }
        personalityPrompt: { type: string }
        tagline: { type: string }
        isLocked: { type: boolean }