# lib/api-zod/src/generated/types/awardXpResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { Character } from "./character";

export interface AwardXpResponse {
  newXp: number;
  unlockedCharacter?: Character | null;
}