# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { CharacterMode } from "./characterMode";

export interface Character {
  id: string;
  name: string;
  mode: CharacterMode;
  xpRequired: number;
  color: string;
  personalityPrompt: string;
  tagline?: string;
  isLocked: boolean;
}