# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  userId: string;
  characterId: string;
  message: string;
  sessionId?: string | null;
  activeTaskId?: string | null;
}