# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { UIBlock } from "./uIBlock";

export interface ChatResponse {
  message: string;
  uiBlocks: UIBlock[];
}