# lib/api-zod/src/generated/types/confirmExtractedBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedTask } from "./extractedTask";

export interface ConfirmExtractedBody {
  userId: string;
  tasks: ExtractedTask[];
}