# lib/api-zod/src/generated/types/extractedTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedTaskResistanceLevel } from "./extractedTaskResistanceLevel";

export interface ExtractedTask {
  title: string;
  durationMinutes?: number | null;
  resistanceLevel: ExtractedTaskResistanceLevel;
  taskType?: string | null;
}