# lib/api-zod/src/generated/types/updateSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSessionBody {
  endTime?: Date | null;
  summary?: string | null;
  activeTaskId?: string | null;
}