# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Spark — ADHD Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyResistanceLevel } from "./updateTaskBodyResistanceLevel";
import type { UpdateTaskBodyStatus } from "./updateTaskBodyStatus";

export interface UpdateTaskBody {
  title?: string;
  status?: UpdateTaskBodyStatus;
  scheduledFor?: Date | null;
  startTime?: Date | null;
  endTime?: Date | null;
  durationMinutes?: number | null;
  resistanceLevel?: UpdateTaskBodyResistanceLevel;
}